Reorganize files in devtools-source-map package
