+++ packages/devtools-source-map/src/util.js
